refactor: restructure screens and add layout
