+++ src/components/Button/styles.ts
@@ -3,8 +3,8 @@ import { transparentize } from "polished";
 import { RectButton } from "react-native-gesture-handler";
 
 export const Container = styled(RectButton)`
-  width: 100%;
   height: 60px;
+  align-self: stretch;
   background: ${({ theme }) => theme.colors.yellow};
   border-radius: 10px;
 
@@ -12,7 +12,7 @@ export const Container = styled(RectButton)`
   align-items: center;
   margin-top: 8px;
 
-  ${({ enabled, theme }) => enabled && css`
+  ${({ enabled, theme }) => !enabled && css`
     background: ${transparentize(0.4, theme.colors.yellow)};
   `}
 `;
