+++ src/components/Input/index.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable no-param-reassign */
 import React, {
   useContext,
   useEffect,
