+++ src/components/Layout/AuthScreenLayout/types.ts
@@ -1,0 +1,7 @@
+export interface AuthScreenProps {
+  title: string;
+  footerButtonText?: string;
+  footerButtonNavigate?: () => void;
+  formBottomButtonText?: string;
+  formBottomButtonNavigate?: () => void;
+}
