+++ src/screens/ForgotPassword/styles.ts
