+++ src/screens/ForgotPassword/types.ts
@@ -1,0 +1,3 @@
+export interface ForgotPasswordFormData {
+  email: string;
+}
