+++ src/screens/SignIn/types.ts
@@ -1,0 +1,4 @@
+export interface SignInFormData {
+  email: string;
+  password: string;
+}
