+++ src/screens/SignUp/types.ts
@@ -1,0 +1,5 @@
+export interface SignUpFormData {
+  name: string;
+  email: string;
+  password: string;
+}
